+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Chg #624: Fix meaning of error message in `OneOf` rule (@arogachev)
 - Chg #625: Improve meaning and use pluralization in error message for `OneOf` and `AtLeast` rules (@arogachev)
 - Chg #626: Disallow `$min` greater than amount of `$attributes` in `AtLeast` configuration (@arogachev)
+- Bug #632: Fix property name usage in error messages of rules in `Nested` rule (@vjik) 
 
 ## 1.1.0 April 06, 2023